// src/app/interface/order.interface.ts
export interface CsvOrderRow {
  orderId: string;
  orderDate: string;
  shippingDate: string;
  qtyShipped: string;
  shipper: string;
  trackAndTraceCode: string;
  trackAndTraceUrl: string;
  customerFirstname: string;
  customerLastname: string;
  shippingAddressStreet: string;
  shippingAddressNumber: string;
  shippingAddressNumberAddition: string;
  shippingAddressPostcode: string;
  shippingAddressCity: string;
  shippingAddressCountry: string;
  customerEmail: string;
  telephone: string;
  sku: string;
  quantity: string;
  EAN: string;
  costPrice: string;
  name: string;
  expectedShippingDate: string;
}